[ScraperError] Move struct into own file
