+++ src/scrapers/ScraperError.cpp
@@ -1,0 +1,5 @@
+#include "scrapers/ScraperError.h"
+
+namespace mediaelch {
+// no-op
+}
